qemu: Implement DTBSurgeon
Update the base QEMU DTB (obtained with -M dumpdtb=<file>), adding or
removing CPU nodes and updating the phandles. Phandle allocation depends
on the number of CPU.

This is fragile: we do not modify the DTB strings block at the moment
(because we would have to reproduce the libfdt algorithm for strings
allocation), so all properties of the template DTB have to be used at
least once in the output DTB, and no new property name can be added.
For the moment it seems good enough. Also, since we can't copy the QEMU
FDT code we only have the output DTB to work with. Good enough for
simple DTBs, but advanced CPU topology, IRQ, SMMU, NUMA and whatnot will
be difficult to match.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>