qemu: Rework gic-version parameter
Support gic-version=host. Normally we can't guess what version the host
supports, but since GICv2 is not supported in Realms, and the probe via
KVM only returns v2 or v3, a successful QEMU invocation with
gic-version=host will necessarily select GICv3.

GICv4 is not supported under KVM, only TCG with EL2, so we don't have to
support it.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>