qemu: Extend -smp parsing
Support topology parameters sockets, clusters, cores and threads for the
-smp parameter.

Mainly we want to avoid throwing an error when a tool like Kata provides
a complex SMP topology. It also generates a more accurate description of
the topology in the device tree, although only for DTB patching. The
simple DTB generation doesn't describe the CPU map (it's not very useful
in a VM without NUMA or cache description.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>