qemu: Support -m maxmem and -m slots
To support memory hotplug QEMU requires reserving RAM addresses with the
maxmem or slots parameters of -m. Specifying a size larger than 255 MB
shifts some high device regions that are located above RAM. To support
this we need to patch the PCI ranges and reg properties.

Remove the check on host IPA bits, because we always assume the QEMU
command is successful, and QEMU itself checks the supported IPA size
from KVM. That means we don't have to pass the IPA size as
realm-measurements parameter anymore, since QEMU calculates the VM IPA
size based on the memory map.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>